member and profile update
